+++ src/components/settings/SettingsDialog.vue
@@ -198,8 +198,8 @@
             calculateSlippage === 'price'
               ? 'Show slippage in $'
               : calculateSlippage === 'bps'
-              ? 'Show slippage in basis point (bps)'
-              : 'Slippage disabled'
+                ? 'Show slippage in basis point (bps)'
+                : 'Slippage disabled'
           "
           v-tippy="{ placement: 'left', boundary: 'window' }"
         >
@@ -721,14 +721,22 @@ export default {
         for (const paneId in this.$store.state.panes.panes) {
           if (/^trades/.test(this.$store.state.panes.panes[paneId].type)) {
             this.$store.dispatch(`${paneId}/generateSwatch`, {
-              buyColor,
-              sellColor,
+              color: buyColor,
+              side: 'buy',
               baseVariance: 0.25
             })
 
+            this.$store.dispatch(`${paneId}/generateSwatch`, {
+              color: sellColor,
+              side: 'sell',
+              baseVariance: 0.25
+            })
+
+            // force refresh
             this.$store.state[paneId].thresholds = JSON.parse(
               JSON.stringify(this.$store.state[paneId].thresholds)
             )
+            this.$store.commit(paneId + '/SET_THRESHOLD_COLOR', {})
           }
         }