+++ src/components/trades/TradesPlaceholder.vue
@@ -15,7 +15,8 @@
         </button>
       </div>
       <button
-        class="mt8 btn -text trades-placeholder__dimmed -small"
+        v-if="paneMarkets.length"
+        class="mt8 btn -text trades-placeholder__dimmed -cases"
         @click="toggleShowMore"
       >
         {{ showMore ? 'Show less' : 'Show more' }}
@@ -95,15 +96,17 @@ export default class TradesPlaceholder extends Vue {
     const minimumLiquidationAmount = this.liquidationsThresholds[0].amount
 
     if (this.showTrades && this.showLiquidations) {
-      return `Waiting for trades or liquidations > ${formatAmount(
+      return `Waiting for trades > ${formatAmount(
         minimumTradeAmount
-      )} / ${formatAmount(minimumLiquidationAmount)}`
+      )} or liquidations > ${formatAmount(minimumLiquidationAmount)}`
     } else if (this.showTrades) {
       return `Waiting for trades > ${formatAmount(minimumTradeAmount)}`
-    } else {
+    } else if (this.showLiquidations) {
       return `Waiting for liquidations > ${formatAmount(
         minimumLiquidationAmount
       )}`
+    } else {
+      return 'Nothing to show, see settings'
     }
   }