+++ src/store/panesSettings/trades.ts
@@ -123,61 +123,43 @@ const actions = {
   generateSwatch(
     { state },
     {
-      buyColor,
-      sellColor,
+      color,
+      side,
       type = 'thresholds',
       baseVariance = 0.15
     }: {
-      buyColor: string
-      sellColor: string
+      color: string
+      side: 'buy' | 'sell'
       type: 'thresholds' | 'liquidations'
       baseVariance: number
     }
   ) {
     const count = state[type].length
     const baseMultipler = (count / 2) * -baseVariance
-    const buyRgb = buyColor ? splitColorCode(buyColor) : null
-    const sellRgb = sellColor ? splitColorCode(sellColor) : null
+    const colorRgb = color ? splitColorCode(color) : null
+    const name = `${side}Color`
+
+    if (!colorRgb) {
+      return
+    }
 
     for (let i = 0; i < count; i++) {
-      if (buyRgb) {
-        if (i === 1) {
-          state[type][i].buyColor = joinRgba([
-            buyRgb[0],
-            buyRgb[1],
-            buyRgb[2],
-            (buyRgb[3] || 1) * 0.8
-          ])
-        } else {
-          const buyScaled = getLogShade(
-            buyRgb,
-            baseMultipler + baseVariance * (i ? i : -0.5)
-          )
-          if (!i) {
-            buyScaled[3] = 0.5
-          }
-          state[type][i].buyColor = joinRgba(buyScaled)
+      if (i === 1) {
+        state[type][i][name] = joinRgba([
+          colorRgb[0],
+          colorRgb[1],
+          colorRgb[2],
+          (colorRgb[3] || 1) * 0.8
+        ])
+      } else {
+        const buyScaled = getLogShade(
+          colorRgb,
+          baseMultipler + baseVariance * (i ? i : -0.5)
+        )
+        if (!i) {
+          buyScaled[3] = 0.5
         }
-      }
-
-      if (sellColor) {
-        if (i === 1) {
-          state[type][i].sellColor = joinRgba([
-            sellRgb[0],
-            sellRgb[1],
-            sellRgb[2],
-            (sellRgb[3] || 1) * 0.8
-          ])
-        } else {
-          const sellScaled = getLogShade(
-            sellRgb,
-            baseMultipler + baseVariance * i
-          )
-          if (!i) {
-            sellScaled[3] = 0.5
-          }
-          state[type][i].sellColor = joinRgba(sellScaled)
-        }
+        state[type][i][name] = joinRgba(buyScaled)
       }
     }
   },
@@ -187,7 +169,7 @@ const actions = {
       data: state,
       type: 'trades-lite'
     })
-  },
+  }
 } as ActionTree<TradesPaneState, ModulesState>
 
 const mutations = {
@@ -280,6 +262,10 @@ const mutations = {
     }
   },
   SET_THRESHOLD_COLOR(state, { id, side, value }) {
+    if (!id) {
+      return
+    }
+
     const threshold = this.getters[state._id + '/getThreshold'](id)
 
     if (threshold) {